Modification du type de la fréquence
Passage de int à float pour la variable de fréquence dans la structure Bpm.
Tentative de modification du programme Processing pour utiliser la date, malheureusement infructueuse.

+++ M4-traitement-des-donnees/headers/donnees.h
@@ -8,7 +8,7 @@
 // recu dans le fichier CSV
 typedef struct {
   int temps;
-  int freq;
+  float freq;
 } Bpm;
 
 // prototype de la fonction de lecture
